Move utils to common directory
