+++ src/commands/sourcemaps/renderer.ts
@@ -2,8 +2,8 @@ import chalk from 'chalk'
 
 import {ICONS} from '../../helpers/formatting'
 import {UploadStatus} from '../../helpers/upload'
+import {pluralize} from './common/utils'
 import {Sourcemap} from './interfaces'
-import {pluralize} from './utils'
 
 export const renderGitWarning = (errorMessage: string) =>
   chalk.yellow(`${ICONS.WARNING} An error occured while invoking git: ${errorMessage}
